Fixed issue with copied symbols not correctly setting the data. Added better handling of background sprites. Lights now move smoothly with the entity they are attached to.

+++ core/src/Roguelike/Lights/Light.java
@@ -17,8 +17,8 @@ public final class Light
 	public float flickerPeriod;
 
 	public float actualIntensity;
-	public int lx;
-	public int ly;
+	public float lx;
+	public float ly;
 
 	private float timeAccumulator;
